+++ addons/bitrix_migration/tests/__init__.py
@@ -6,3 +6,5 @@ from . import test_employee_avatars
 from . import test_meetings
 from . import test_meeting_comments
 from . import test_meeting_attachments
+from . import test_fired_employees
+from . import test_purge_noise